fix: restructure-file-system
